Some searching changes
