+++ src/main/java/me/canelex/spidey/objects/search/GoogleSearch.java
@@ -2,22 +2,44 @@ package me.canelex.spidey.objects.search;
 
 import me.canelex.spidey.Secrets;
 import me.canelex.spidey.utils.Utils;
+import org.apache.commons.text.StringEscapeUtils;
+
+import java.net.URLDecoder;
+import java.nio.charset.StandardCharsets;
 
 public class GoogleSearch
 {
-    private GoogleSearch()
-    {
-        super();
-    }
+    private String title;
+    private String content;
+    private String url;
+    private static final String GOOGLE_URL = "https://www.googleapis.com/customsearch/v1/?cx=015021391643023377625:kq7ex3xgvoq&key=" + Secrets.GOOGLE_API_KEY + "&num=1&q=%s";
 
-    private static final String GOOGLE_URL = "https://www.googleapis.com/customsearch/v1/?cx=%s&key=%s&num=1&q=%s";
-    private static final String GOOGLE_API_KEY = Secrets.GOOGLE_API_KEY;
-
-    public static SearchResult performSearch(String terms)
+    public final GoogleSearch getResult(String terms)
     {
         terms = terms.replace(" ", "%20");
-        final var searchUrl = String.format(GOOGLE_URL, "015021391643023377625:kq7ex3xgvoq", GOOGLE_API_KEY, terms);
+        final var searchUrl = String.format(GOOGLE_URL, terms);
         final var o = Utils.getJson(searchUrl).getArray("items").getObject(0);
-        return SearchResult.fromGoogle(o);
+        this.title = cleanString(o.getString("title"));
+        this.content = cleanString(o.getString("snippet"));
+        this.url = URLDecoder.decode(cleanString(o.getString("link")), StandardCharsets.UTF_8);
+        return this;
+    }
+
+    public final String getContent()
+    {
+        if (url.startsWith("https://www.youtube.com/watch?"))
+            return "<" + url.replace("https://www.youtube.com/watch?v=", "https://youtu.be/") + "> - *" + title + "*: \"" + content + "\"";
+        else
+            return "<" + url.replace("www.", "") + "> - *" + title + "*: \"" + content + "\"";
+    }
+
+    private static String cleanString(String uncleanString)
+    {
+        return StringEscapeUtils.unescapeJava(
+                StringEscapeUtils.unescapeHtml4(
+                        uncleanString
+                                .replaceAll("\\s+", " ")
+                                .replaceAll("<.*?>", "")
+                                .replaceAll("\"", "")));
     }
 }
